Avance tesis

+++ hale-jobs-api/src/main/java/com/services/chambitas/domain/dto/UserCVBasicDTO.java
@@ -11,6 +11,6 @@ public class UserCVBasicDTO {
 	
 	private String jobTitle;
 	
-	private double salary;
+	private Long salary;
 
 }
